Predictor Alquiler Definitivo/requirements.txt
sqlalchemy
bcrypt
python-dotenv
streamlit
pandas
numpy
plotly
pydeck
psycopg2
scikit-learn
beautifulsoup4
undetected-chromedriver
geopy
